Remove unused bits from .gitignore
I'm taking the liberty to edit your `.gitignore` files and remove the unnecessary entries (most of them). It's not a problem to have a big `.gitignore` file, but it's kind of making the diff hard to read for the pull request. You don't need these entries anyway, so I'm removing them.